refactor: Class 전체 문제 목록 폴더 분리
