+++ src/pages/Class/AllProblems/hooks/query/useAllProblemListQuery.ts
@@ -4,16 +4,16 @@ import { AxiosError } from 'axios';
 import { useSuspenseQuery } from '@/hooks/useSuspenseQuery';
 import { QUERY_KEYS } from '@/constants';
 
-import { getProblemList } from '../../api';
+import { getAllProblemList } from '../../api';
 
-export const useProblemListQuery = (
+export const useAllProblemListQuery = (
   { keyword }: { keyword: string },
   options?: UseQueryOptions<ProblemListResponse, AxiosError, ProblemListResponse, [string, string]>
 ) => {
   return useSuspenseQuery(
     [QUERY_KEYS.CLASS_PROBLEM, keyword],
     async ({ queryKey: [, keyword] }) => {
-      const { data } = await getProblemList({ keyword });
+      const { data } = await getAllProblemList({ keyword });
       return data;
     },
     {
