+++ src/pages/Class/AllProblems/index.ts
@@ -1,0 +1,2 @@
+export * from './AllProblemList';
+export * from './AllProblemDetail';
