+++ src/pages/Class/Problem/api/index.ts
@@ -2,28 +2,12 @@ import { api, fileApi } from '@/api';
 
 const API_URL = `/problems`;
 
-const getProblemList = ({ keyword }: { keyword: string }) => {
-  return api.get(`${API_URL}?keyword=${keyword}`);
-};
-
-const getProblem = (problemId: string) => {
-  return api.get(`${API_URL}/${problemId}`);
-};
-
 const createProblem = (payload: FormData) => {
   return fileApi.post(`${API_URL}/`, payload);
 };
 
 const editProblem = (problemId: string, payload: FormData) => {
   return fileApi.put(`${API_URL}/${problemId}/`, payload);
-};
-
-const editProblemPublic = (problemId: string) => {
-  return api.post(`${API_URL}/${problemId}/check/`);
-};
-
-const deleteProblem = (problemId: string) => {
-  return api.delete(`${API_URL}/${problemId}`);
 };
 
 const getContestProblem = ({ classId, contestId, contestProblemId }: ContestProblemRequest) => {
@@ -60,12 +44,8 @@ const createContestProblemSumbissionCheck = ({
 };
 
 export {
-  getProblemList,
-  getProblem,
   createProblem,
   editProblem,
-  editProblemPublic,
-  deleteProblem,
   getContestProblem,
   getContestProblemSubmission,
   createContestProblemSubmission,
