+++ src/pages/Class/Problem/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './query';
-export * from './useClassProblemListTable';
