+++ src/pages/Class/Problem/hooks/query/index.ts
@@ -1,9 +1,5 @@
-export * from './useProblemQuery';
 export * from './useCreateProblemMutation';
 export * from './useEditProblemMutation';
-export * from './useDeleteProblemMutation';
-export * from './useEditProblemPublicMutation';
-export * from './useProblemListQuery';
 
 export * from './useContestProblemQuery';
 export * from './useContestProblemSubmissionQuery';
