+++ src/pages/Class/Problem/index.ts
@@ -1,11 +1,9 @@
 export * from './ProblemCreate';
 export * from './ProblemEdit';
 
-export * from './AllProblemDetail';
 export * from './ContestProblemDetail';
 
 export * from './ProblemDescription';
 export * from './ProblemData';
 export * from './ProblemLeaderBoard';
 export * from './ProblemSubmission';
-export * from './ClassProblemList';
