+++ src/pages/Class/index.ts
@@ -1,7 +1,9 @@
 export * from './ClassList';
 export * from './ClassListEdit';
 export * from './Class';
-export * from './Contest';
+
+export * from './AllProblems';
 export * from './StudentManagement';
+export * from './Contest';
 
 export * from './Problem';
